pyproject.toml
[tool.poetry]
name = "domen"
version = "0.1.0"
description = ""
authors = ["'Horrorzero' <vladpushkar82@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.0.0b8"
asyncwhois = "^1.0.7"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
